[#360] - Traslado de rutas a archivo app.routes.ts
