+++ src/app/components/story-navigation-bar/story-navigation-bar.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnChanges, SimpleChanges } from '@angular/core';
 import { Publication, Storylist } from '@models/storylist.model';
 import { Story } from '@models/story.model';
-import { APP_ROUTE_TREE } from '../../app-routing.module';
+import { APP_ROUTE_TREE } from '../../app.routes';
 import { NgxSkeletonLoaderModule } from 'ngx-skeleton-loader';
 import { StoryEditionDateLabelComponent } from '../story-edition-date-label/story-edition-date-label.component';
 import { RouterLink } from '@angular/router';
