+++ src/main.server.ts
@@ -1,2 +1,3 @@
 
-export { AppServerModule } from './app/app.server.module';
+export { AppServerModule } from './app/app.server.module';  
+
